refactor(context): split ThemeContext and ThemeProvider to fix Fast Refresh warning and remove unused App.css
