+++ .gitignore
@@ -22,3 +22,8 @@ dist-ssr
 *.njsproj
 *.sln
 *.sw?
+
+# Antigravity CLI metadata
+.antigravitycli/
+.antigravity/
+
