+++ src/context/ThemeContext.js
@@ -1,0 +1,11 @@
+import { createContext, useContext } from "react";
+
+export const ThemeContext = createContext(undefined);
+
+export const useThemeContext = () => {
+  const context = useContext(ThemeContext);
+  if (context === undefined) {
+    throw new Error("useThemeContext must be used within a ThemeProvider");
+  }
+  return context;
+};
